+++ jgclark.Reviews/src/allProjectsListHelpers.js
@@ -4,7 +4,7 @@
 //-----------------------------------------------------------------------------
 // Supporting functions that deal with the allProjects list.
 // by @jgclark
-// Last updated 2026-05-11 for v2.0.0.b32 by @CursorAI
+// Last updated 2026-07-17 for v2.0.6 by @CursorAI
 //-----------------------------------------------------------------------------
 
 import moment from 'moment/min/moment-with-locales'