+++ jgclark.Reviews/src/index.js
@@ -3,7 +3,7 @@
 //-----------------------------------------------------------------------------
 // Index for Reviews plugin
 // by Jonathan Clark
-// Last updated 2026-07-17 for v2.0.2 by @jgclark + @CursorAI
+// Last updated 2026-07-17 for v2.0.6 by @jgclark + @CursorAI
 //-----------------------------------------------------------------------------
 
 // allow changes in plugin.json to trigger recompilation